refactor: change double to float for improved performance in multiple functions and update experfc signature

+++ include/qalgorithms_qpeaks.h
@@ -181,7 +181,7 @@ namespace qAlgorithms
      * @param scale
      * @return float
      */
-    inline double multiplyVecMatrixVecTranspose(const double vec[4],
+    inline float multiplyVecMatrixVecTranspose(const float vec[4],
                                                 size_t scale);
 
     float apexToEdgeRatio(
